Move StickyFaultsProcessor and rename
